ir_recv_emit_controller/server/requirements.txt
Flask
Flask-SocketIO
pyserial
dotenv